+++ server/middleware/authentication/user.js
@@ -1,3 +1,6 @@
+// user.js contains the middleware for making sure a user is properly authenticated
+
+
 var bcrypt = require('bcryptjs');
 var moment = require('moment');
 var jwt = require('jwt-simple');
@@ -6,6 +9,8 @@ var validator = require('validator');
 var status = require('../../config/status.js');
 
 module.exports = {
+  
+  // ensure either authorization header or authorization url-parameter contains a valid token
   ensureAuthenticated: function(req, res, next){
     var token_alt_0 = null;
     var token_alt_1 = null;
@@ -15,13 +20,14 @@ module.exports = {
     if (req.params.authorization) {
       token_alt_1 = req.params.authorization;
     }
+    // ensure either parameter or header exists
     if (!token_alt_0 && !token_alt_1) {
       return res.status(401).send(help.sendError(req.language, 'USER_NOT_FOUND'));
     }
 
     var token = token_alt_0 || token_alt_1;
 
-
+    // check if given token is actually valid
     var payload = null;
     try {
       payload = jwt.decode(token, config.TOKEN_SECRET);
@@ -33,6 +39,7 @@ module.exports = {
     if (payload.exp <= moment().unix()) {
       return res.status(401).send(help.sendError(req.language, 'USER_NOT_FOUND'));
     }
+    // set user-id to be used in the endpoint-handler
     req.user = payload.sub;
     next();
   },
