+++ server/middleware/index.js
@@ -1,3 +1,6 @@
+// the parameterValidation-files validates the parameters for the endpoints "contained" in that category
+
+
 module.exports = {
   misc: {
     language: require('./misc/language.js')
