+++ server/routes/auth.js
@@ -12,12 +12,14 @@ var bcrypt = require('bcryptjs');
 
 module.exports = function(app) {
 
+  // log in user and return authentication token
   app.post('/api/auth/login', Middleware.misc.language, Middleware.parameterValidation.auth.login, function(req, res) {
     User.findOne({ email: req.body.email }, function(err, user) {
       if (err) {throw err;}
       if (!user) {
         return res.status(406).send(help.sendError(req.language, 'USER_DOES_NOT_EXIST'));
       }
+      // check if password is matching saved one
       bcrypt.compare(req.body.password, user.password, function(err, isMatch) {
         if (err) {throw err;}
         if (!isMatch) {
@@ -28,12 +30,14 @@ module.exports = function(app) {
     })
   })
 
+  // sign up user and return authentication token
   app.put('/api/auth/signup', Middleware.misc.language, Middleware.parameterValidation.auth.signup, function(req, res) {
     User.findOne({ email: req.body.email }, function(err, user) {
       if (err) {throw err;}
       if (user) {
         return res.status(406).send(help.sendError(req.language, 'USER_ALREADY_EXISTS'));
       }
+      // initialize new user
       var newUser = new User({
         first_name: req.body.first_name,
         last_name: req.body.last_name,
